requirements.txt
numpy
pandas
gensim
nltk
sklearn
sentence-transformers